+++ examples/05-mesh-optimizer/mesh-optimizer.cpp
@@ -168,7 +168,7 @@ void Run()
         RNDR_TRACE_END(Process_events);
 
         const float ratio = static_cast<Rndr::real>(window.GetWidth()) / static_cast<Rndr::real>(window.GetHeight());
-        const float angle = std::fmod(10 * Rndr::GetSystemTime(), 360.0f);
+        const float angle = static_cast<float>(std::fmod(10 * Rndr::GetSystemTime(), 360.0));
         const math::Transform t1 =
             math::Translate(math::Vector3(-0.5f, -0.5f, -1.5f)) * math::Rotate(angle, math::Vector3(0.0f, 1.0f, 0.0f)) * math::RotateX(-90);
         const math::Transform t2 =
